// src/app/api/rss/feeds.ts
// This file is auto-generated for sharing feeds between API and frontend
const feeds = [
  { journalName: 'IEEE Transactions on Eng. Management', url: 'https://ieeexplore.ieee.org/rss/TOC17.XML', type: 'standard' },
  { journalName: 'Int. Journal of Technology Management', url: 'https://www.inderscienceonline.com/action/showFeed?type=etoc&feed=rss&jc=ijtm', type: 'standard' },
  { journalName: 'Int. Journal of Innovation Management', url: 'https://www.worldscientific.com/action/showFeed?type=etoc&feed=rss&jc=ijim', type: 'standard' },
  { journalName: 'Computers in Industry', url: 'https://rss.sciencedirect.com/publication/science/01663615', type: 'sciencedirect' },
  { journalName: 'Technovation, Technological Innovation', url: 'https://rss.sciencedirect.com/publication/science/01664972', type: 'sciencedirect' },
  { journalName: 'Computers & Industrial Engineering', url: 'https://rss.sciencedirect.com/publication/science/03608352', type: 'sciencedirect' },
  { journalName: 'Advanced Engineering Informatics', url: 'https://rss.sciencedirect.com/publication/science/14740346', type: 'sciencedirect' },
  { journalName: 'Technological Forecasting & Social Change', url: 'https://rss.sciencedirect.com/publication/science/00401625', type: 'sciencedirect' },
  { journalName: 'Technology in Society', url: 'https://rss.sciencedirect.com/publication/science/0160791X', type: 'sciencedirect' },
  { journalName: 'World Patent Information', url: 'https://rss.sciencedirect.com/publication/science/01722190', type: 'sciencedirect' },
];
export default feeds;